fix: prevent premature exit after an assertion failure
