+++ src/modules/essentials/assert.ts
@@ -1,4 +1,4 @@
 import nodeAssert from 'node:assert';
-import { createAssert } from '../../services/assert.js';
+import { createAssert } from '../builders/assert.js';
 
 export const assert = createAssert(nodeAssert);
