+++ src/modules/essentials/strict.ts
@@ -1,4 +1,4 @@
 import nodeAssert from 'node:assert/strict';
-import { createAssert } from '../../services/assert.js';
+import { createAssert } from '../builders/assert.js';
 
 export const strict = createAssert(nodeAssert);
