+++ src/services/assert.ts
@@ -1,12 +1,10 @@
 import type { ProcessAssertionOptions } from '../@types/assert.js';
-import type assert from 'node:assert';
-import type { AssertPredicate } from 'node:assert';
 import { AssertionError } from 'node:assert';
-import { cwd as processCWD, env, exit } from 'node:process';
+import process, { cwd as processCWD, env } from 'node:process';
 import path from 'node:path';
 import { findFile } from '../parsers/find-file-from-stack.js';
 import { parseResultType } from '../parsers/assert.js';
-import { nodeVersion } from '../parsers/get-runtime.js';
+
 import { indentation } from '../configs/indentation.js';
 import { format } from './format.js';
 import { Write } from './write.js';
@@ -14,30 +12,22 @@ import { Write } from './write.js';
 const cwd = processCWD();
 const regexFile = /file:(\/\/)?/;
 
-export const processAssert = async (
-  cb: () => void | Promise<void>,
-  options: ProcessAssertionOptions
-) => {
+const assertProcessor = () => {
   const isPoku = typeof env?.FILE === 'string' && env?.FILE.length > 0;
   const FILE = env.FILE;
+
   let preIdentation = '';
 
-  if (indentation.hasDescribe) {
-    preIdentation += '  ';
-  }
-
-  if (indentation.hasItOrTest) {
-    preIdentation += '  ';
-  }
-
-  try {
-    const cbResult = cb();
-
-    if (cbResult instanceof Promise) {
-      await cbResult;
-    }
-
+  const handleSuccess = (options: ProcessAssertionOptions) => {
     if (typeof options.message === 'string') {
+      if (indentation.hasDescribe) {
+        preIdentation += '  ';
+      }
+
+      if (indentation.hasItOrTest) {
+        preIdentation += '  ';
+      }
+
       const message =
         isPoku && !indentation.hasDescribe && !indentation.hasItOrTest
           ? `${preIdentation}${format(`${format(`✔ ${options.message}`).bold()} ${format(`› ${FILE}`).success().dim()}`).success()}`
@@ -45,11 +35,25 @@ export const processAssert = async (
 
       Write.log(message);
     }
-  } catch (error) {
+
+    preIdentation = '';
+  };
+
+  const handleError = (error: unknown, options: ProcessAssertionOptions) => {
+    process.exitCode = 1;
+
     if (error instanceof AssertionError) {
       const { code, actual, expected, operator } = error;
       const absolutePath = findFile(error).replace(regexFile, '');
       const file = path.relative(path.resolve(cwd), absolutePath);
+
+      if (indentation.hasDescribe) {
+        preIdentation += '  ';
+      }
+
+      if (indentation.hasItOrTest) {
+        preIdentation += '  ';
+      }
 
       let message = '';
 
@@ -96,6 +100,8 @@ export const processAssert = async (
         for (const line of splitExpected) {
           Write.log(`${preIdentation}  ${format(line).success().bold()}`);
         }
+
+        preIdentation = '';
       }
 
       if (options.throw) {
@@ -103,340 +109,37 @@ export const processAssert = async (
         Write.hr();
       }
 
-      exit(1);
+      if (isPoku) {
+        throw error;
+      }
     }
 
-    /* c8 ignore next 2 */ // Unknown external error
+    /* c8 ignore next */ // Unknown external error
     throw error;
-  }
+  };
+
+  const processAssert = (cb: () => void, options: ProcessAssertionOptions) => {
+    try {
+      cb();
+      handleSuccess(options);
+    } catch (error) {
+      handleError(error, options);
+    }
+  };
+
+  const processAsyncAssert = async (
+    cb: () => Promise<void>,
+    options: ProcessAssertionOptions
+  ) => {
+    try {
+      await cb();
+      handleSuccess(options);
+    } catch (error) {
+      handleError(error, options);
+    }
+  };
+
+  return { processAssert, processAsyncAssert };
 };
 
-export const createAssert = (nodeAssert: typeof assert) => {
-  const ok = (
-    value: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(
-      () => {
-        nodeAssert.ok(value);
-      },
-      { message }
-    );
-  };
-
-  const equal = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(
-      () => {
-        nodeAssert.equal(actual, expected);
-      },
-      { message }
-    );
-  };
-
-  const deepEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.deepEqual(actual, expected), { message });
-  };
-
-  const strictEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.strictEqual(actual, expected), { message });
-  };
-
-  const deepStrictEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.deepStrictEqual(actual, expected), {
-      message,
-    });
-  };
-
-  const notEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.notEqual(actual, expected), {
-      message,
-    });
-  };
-
-  const notDeepEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.notDeepEqual(actual, expected), { message });
-  };
-
-  const notStrictEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.notStrictEqual(actual, expected), {
-      message,
-    });
-  };
-
-  const notDeepStrictEqual = (
-    actual: unknown,
-    expected: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(() => nodeAssert.notDeepStrictEqual(actual, expected), {
-      message,
-    });
-  };
-
-  const ifError = (
-    value: unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    processAssert(
-      () => {
-        nodeAssert.ifError(value);
-      },
-      {
-        message,
-        defaultMessage: 'Expected no error, but received an error',
-        hideDiff: true,
-        throw: true,
-      }
-    );
-  };
-
-  const fail = (message?: ProcessAssertionOptions['message']): never => {
-    processAssert(
-      () => {
-        nodeAssert.fail(message);
-      },
-      {
-        message,
-        defaultMessage: 'Test failed intentionally',
-        hideDiff: true,
-      }
-    );
-
-    process.exit(1);
-  };
-
-  function doesNotThrow(
-    block: () => unknown,
-    message?: string | ProcessAssertionOptions['message']
-  ): void;
-  function doesNotThrow(
-    block: () => unknown,
-    error: AssertPredicate,
-    message?: ProcessAssertionOptions['message']
-  ): void;
-  function doesNotThrow(
-    block: () => unknown,
-    errorOrMessage?: AssertPredicate | ProcessAssertionOptions['message'],
-    message?: ProcessAssertionOptions['message']
-  ): void {
-    processAssert(
-      () => {
-        if (
-          typeof errorOrMessage === 'function' ||
-          errorOrMessage instanceof RegExp ||
-          typeof errorOrMessage === 'object'
-        ) {
-          nodeAssert.doesNotThrow(block, errorOrMessage, message);
-        } else {
-          const msg =
-            typeof errorOrMessage === 'string' ? errorOrMessage : message;
-          nodeAssert.doesNotThrow(block, msg);
-        }
-      },
-      {
-        message: typeof errorOrMessage === 'string' ? errorOrMessage : message,
-        defaultMessage: 'Expected function not to throw',
-        hideDiff: true,
-        throw: true,
-      }
-    );
-  }
-
-  function throws(
-    block: () => unknown,
-    message?: ProcessAssertionOptions['message']
-  ): void;
-  function throws(
-    block: () => unknown,
-    error: AssertPredicate,
-    message?: ProcessAssertionOptions['message']
-  ): void;
-  function throws(
-    block: () => unknown,
-    errorOrMessage?: AssertPredicate | ProcessAssertionOptions['message'],
-    message?: ProcessAssertionOptions['message']
-  ): void {
-    if (
-      typeof errorOrMessage === 'function' ||
-      errorOrMessage instanceof RegExp ||
-      typeof errorOrMessage === 'object'
-    ) {
-      processAssert(() => nodeAssert.throws(block, errorOrMessage), {
-        message,
-        defaultMessage: 'Expected function to throw',
-        hideDiff: true,
-      });
-    } else {
-      const msg =
-        typeof errorOrMessage !== 'undefined' ? errorOrMessage : message;
-
-      processAssert(() => nodeAssert.throws(block, message), {
-        message: msg,
-        defaultMessage: 'Expected function to throw',
-        hideDiff: true,
-      });
-    }
-  }
-
-  function rejects(
-    block: (() => Promise<unknown>) | Promise<unknown>,
-    message?: ProcessAssertionOptions['message']
-  ): Promise<void>;
-  function rejects(
-    block: (() => Promise<unknown>) | Promise<unknown>,
-    error: AssertPredicate,
-    message?: ProcessAssertionOptions['message']
-  ): Promise<void>;
-  async function rejects(
-    block: (() => Promise<unknown>) | Promise<unknown>,
-    errorOrMessage?: AssertPredicate | ProcessAssertionOptions['message'],
-    message?: ProcessAssertionOptions['message']
-  ): Promise<void> {
-    await processAssert(
-      async () => {
-        if (
-          typeof errorOrMessage === 'function' ||
-          errorOrMessage instanceof RegExp ||
-          typeof errorOrMessage === 'object'
-        ) {
-          await nodeAssert.rejects(block, errorOrMessage, message);
-        } else {
-          const msg =
-            typeof errorOrMessage === 'string' ? errorOrMessage : message;
-          await nodeAssert.rejects(block, msg);
-        }
-      },
-      {
-        message: typeof errorOrMessage === 'string' ? errorOrMessage : message,
-        defaultMessage: 'Expected promise to be rejected with specified error',
-        hideDiff: true,
-        throw: true,
-      }
-    );
-  }
-
-  function doesNotReject(
-    block: (() => Promise<unknown>) | Promise<unknown>,
-    message?: ProcessAssertionOptions['message']
-  ): Promise<void>;
-  function doesNotReject(
-    block: (() => Promise<unknown>) | Promise<unknown>,
-    error: AssertPredicate,
-    message?: ProcessAssertionOptions['message']
-  ): Promise<void>;
-  async function doesNotReject(
-    block: (() => Promise<unknown>) | Promise<unknown>,
-    errorOrMessage?: AssertPredicate | ProcessAssertionOptions['message'],
-    message?: ProcessAssertionOptions['message']
-  ): Promise<void> {
-    await processAssert(
-      async () => {
-        if (
-          typeof errorOrMessage === 'function' ||
-          errorOrMessage instanceof RegExp ||
-          typeof errorOrMessage === 'object'
-        ) {
-          await nodeAssert.doesNotReject(block, errorOrMessage, message);
-        } else {
-          await nodeAssert.doesNotReject(block, message);
-        }
-      },
-      {
-        message: typeof errorOrMessage === 'string' ? errorOrMessage : message,
-        defaultMessage: 'Got unwanted rejection',
-        hideDiff: true,
-        throw: true,
-      }
-    );
-  }
-
-  const match = (
-    value: string,
-    regExp: RegExp,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    /* c8 ignore next 3 */ // Platform version
-    if (typeof nodeVersion === 'number' && nodeVersion < 12) {
-      throw new Error('match is available from Node.js 12 or higher');
-    }
-
-    processAssert(() => nodeAssert?.match(value, regExp), {
-      message,
-      actual: 'Value',
-      expected: 'RegExp',
-      defaultMessage: 'Value should match regExp',
-    });
-  };
-
-  const doesNotMatch = (
-    value: string,
-    regExp: RegExp,
-    message?: ProcessAssertionOptions['message']
-  ): void => {
-    /* c8 ignore next 3 */ // Platform version
-    if (typeof nodeVersion === 'number' && nodeVersion < 12) {
-      throw new Error('doesNotMatch is available from Node.js 12 or higher');
-    }
-
-    processAssert(() => nodeAssert.doesNotMatch(value, regExp), {
-      message,
-      actual: 'Value',
-      expected: 'RegExp',
-      defaultMessage: 'Value should not match regExp',
-    });
-  };
-
-  const assert = Object.assign(
-    (value: unknown, message?: ProcessAssertionOptions['message']) =>
-      ok(value, message),
-    {
-      ok,
-      equal,
-      deepEqual,
-      strictEqual,
-      deepStrictEqual,
-      doesNotMatch,
-      doesNotReject,
-      throws,
-      doesNotThrow,
-      notEqual,
-      notDeepEqual,
-      notStrictEqual,
-      notDeepStrictEqual,
-      match,
-      ifError,
-      fail,
-      rejects,
-    }
-  );
-
-  return assert;
-};
+export const { processAssert, processAsyncAssert } = assertProcessor();
