+++ src/services/write.ts
@@ -6,7 +6,7 @@ export const Write = {
     stdout.write(`${String(data)}\n`);
   },
   hr: () => {
-    const line = '⎯'.repeat(stdout.columns - 10 || 40);
+    const line = '─'.repeat(stdout.columns - 10 || 40);
 
     Write.log(`\n\x1b[2m\x1b[90m${line}\x1b[0m\n`);
   },
